+++ libs/core/CHANGELOG.md
@@ -1,3 +1,63 @@
+## 0.3.8 (2025-11-24)
+
+### 🩹 Fixes
+
+- build libs for es2022 and commonjs ([7a7c037](https://github.com/golemui/formforge/commit/7a7c037))
+
+### ❤️ Thank You
+
+- Raul Jimenez @Elecash
+
+## 0.3.7 (2025-11-23)
+
+This was a version bump only for core to align it with other projects, there were no code changes.
+
+## 0.3.6 (2025-11-23)
+
+### 🩹 Fixes
+
+- **react:** fix react FieldRenderer ([0cb7882](https://github.com/golemui/formforge/commit/0cb7882))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
+## 0.3.5 (2025-11-23)
+
+This was a version bump only for core to align it with other projects, there were no code changes.
+
+## 0.3.4 (2025-11-23)
+
+This was a version bump only for core to align it with other projects, there were no code changes.
+
+## 0.3.3 (2025-11-23)
+
+This was a version bump only for core to align it with other projects, there were no code changes.
+
+## 0.3.2 (2025-11-23)
+
+### 🩹 Fixes
+
+- **react:** fix react context provider duplicity ([dcad867](https://github.com/golemui/formforge/commit/dcad867))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
+## 0.3.1 (2025-11-23)
+
+This was a version bump only for core to align it with other projects, there were no code changes.
+
+## 0.3.0 (2025-11-23)
+
+### 🚀 Features
+
+- **validators:** Make validators pluggable ([397e67c](https://github.com/golemui/formforge/commit/397e67c))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
 ## 0.2.4 (2025-11-22)
 
 This was a version bump only for core to align it with other projects, there were no code changes.