+++ libs/react/CHANGELOG.md
@@ -1,3 +1,88 @@
+## 0.3.8 (2025-11-24)
+
+### 🩹 Fixes
+
+- build libs for es2022 and commonjs ([7a7c037](https://github.com/golemui/formforge/commit/7a7c037))
+
+### ❤️ Thank You
+
+- Raul Jimenez @Elecash
+
+## 0.3.7 (2025-11-23)
+
+### 🩹 Fixes
+
+- disable minification for testing purposes ([eee4d87](https://github.com/golemui/formforge/commit/eee4d87))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
+## 0.3.6 (2025-11-23)
+
+### 🩹 Fixes
+
+- **react:** fix react FieldRenderer ([e29d12e](https://github.com/golemui/formforge/commit/e29d12e))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
+## 0.3.5 (2025-11-23)
+
+### 🩹 Fixes
+
+- **react:** fix react FieldRenderer ([2ff9d71](https://github.com/golemui/formforge/commit/2ff9d71))
+- **react:** remove the need to pass formContext to components ([ddce6aa](https://github.com/golemui/formforge/commit/ddce6aa))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
+## 0.3.4 (2025-11-23)
+
+### 🩹 Fixes
+
+- **react:** fix react FieldRenderer ([57be08d](https://github.com/golemui/formforge/commit/57be08d))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
+## 0.3.3 (2025-11-23)
+
+### 🩹 Fixes
+
+- **react:** fix react context provider null value ([ee9505f](https://github.com/golemui/formforge/commit/ee9505f))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
+## 0.3.2 (2025-11-23)
+
+### 🩹 Fixes
+
+- **react:** fix react context provider duplicity ([dcad867](https://github.com/golemui/formforge/commit/dcad867))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
+## 0.3.1 (2025-11-23)
+
+This was a version bump only for react to align it with other projects, there were no code changes.
+
+## 0.3.0 (2025-11-23)
+
+### 🚀 Features
+
+- **validators:** Make validators pluggable ([397e67c](https://github.com/golemui/formforge/commit/397e67c))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
 ## 0.2.4 (2025-11-22)
 
 ### 🩹 Fixes